Refactor: Splash 수정
